# apps/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Tanban Restaurant SaaS API
  version: 0.1.0
  description: 所有业务时间均按北京时间（Asia/Shanghai，UTC+8）输入、计算和展示；响应中的本地日期时间统一为 `YYYY-MM-DD HH:mm:ss`。需表达绝对时刻的请求可使用带 `+08:00` 的 RFC3339。
servers:
  - url: https://tbapi.666qwe.cn/api/v1
paths:
...
  /merchant/media-assets/upload:
    post:
      summary: 上传商户装修图片
      description: 图片按当前商户和门店隔离存储，支持 JPEG、PNG 和 GIF，单文件最大 8 MiB。
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: {type: string, format: binary}
                name: {type: string, maxLength: 120}
                group_id: {type: integer, format: int64, minimum: 0}
      responses:
        '201':
          description: Created media asset
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: {$ref: '#/components/schemas/MediaAsset'}
        '400': {description: Invalid multipart form or name}
        '413': {description: File exceeds 8 MiB}
        '415': {description: Unsupported or invalid image}
...
  /merchant/marketing/apps:
    get:
      summary: 已开放营销应用及活动状态汇总
      security: [{bearerAuth: []}]
      responses: {'200': {description: Marketing application registry}}
...
  /public/media/{storagePath}:
    get:
      summary: 读取已上传的公开图片
      description: 只读取处于 ACTIVE 状态的图片；storagePath 包含服务端生成的多段路径。
      security: []
      parameters:
        - in: path
          name: storagePath
          required: true
          schema: {type: string}
      responses:
        '200': {description: Raw immutable image bytes}
        '404': {description: Media asset is unavailable}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    MediaAsset:
      type: object
      required: [id, name, kind, url, storageKey, mimeType, width, height, sizeBytes, status, createdAt, group_name]
      properties:
        id: {type: integer, format: int64}
        name: {type: string}
        kind: {type: string, enum: [IMAGE]}
        url: {type: string, format: uri}
        storageKey: {type: string}
        mimeType: {type: string, enum: [image/jpeg, image/png, image/gif]}
        width: {type: integer}
        height: {type: integer}
        sizeBytes: {type: integer, format: int64}
        status: {type: string, enum: [ACTIVE]}
        createdAt: {type: string, format: date-time}
        group_id: {type: integer, format: int64, nullable: true}
        group_name: {type: string}